requirements.txt
flask
flask_cors
youtube-transcript-api
python-dotenv